third_party/blink/web_tests/svg/dynamic-updates/SVGRectElement-svgdom-y-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'y' property of the SVGRectElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rectElement.y.baseVal.value is -150
PASS rectElement.y.baseVal.value is 0
PASS successfullyParsed is true

TEST COMPLETE
